Fix MsgGetHomingAttackPriority field type.

+++ Sonic/Message/MsgGetHomingAttackPriority.h
@@ -9,10 +9,10 @@ namespace Sonic::Message
     public:
         HH_FND_MSG_MAKE_TYPE(0x167FAA8);
 
-        size_t* m_pPriority;
+        int32_t* m_pPriority;
 
-        MsgGetHomingAttackPriority(size_t* in_pPriority) : m_pPriority(in_pPriority) {}
-        MsgGetHomingAttackPriority(size_t& in_rPriority) : m_pPriority(&in_rPriority) {}
+        MsgGetHomingAttackPriority(int32_t* in_pPriority) : m_pPriority(in_pPriority) {}
+        MsgGetHomingAttackPriority(int32_t& in_rPriority) : m_pPriority(&in_rPriority) {}
     };
 
     BB_ASSERT_OFFSETOF(MsgGetHomingAttackPriority, m_pPriority, 0x10);
